requirements.txt
python-telegram-bot[all]>=21.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
openai>=1.0.0
